# .legacy/client/docs/swagger-ui-legacy/swagger.yaml
swagger: '2.0'
info:
  title: xco modules - legacy endpoints docs
  description: A REST interface for state queries
  version: 1.0.0
host: testnet.xco.world
basePath: /rest
schemes:
  - https
paths:
...
  '/project/{project_did}':
    get:
      operationId: ProjectDoc
      description: Get a project document by its project DID
      responses:
        '200':
          description: A successful response.
          schema:
            type: object
            properties:
              txHash:
                type: string
              projectDid:
                type: string
              senderDid:
                type: string
              pubKey:
                type: string
              data:
                type: string
                format: byte
        default:
          description: An unexpected error response
          schema:
            type: object
            properties:
              error:
                type: string
              code:
                type: integer
                format: int32
              message:
                type: string
              details:
                type: array
                items:
                  type: object
                  properties:
                    type_url:
                      type: string
                    value:
                      type: string
                      format: byte
      parameters:
        - name: project_did
          in: path
          required: true
          type: string
      tags:
        - Project
  '/projectAccounts/{project_did}':
    get:
      operationId: ProjectAccounts
      description: Get a specific project's list of project accounts
      responses:
        '200':
          description: A successful response.
          schema:
            type: object
            properties:
              map:
                type: object
                additionalProperties:
                  type: string
        default:
          description: An unexpected error response
          schema:
            type: object
            properties:
              error:
                type: string
              code:
                type: integer
                format: int32
              message:
                type: string
              details:
                type: array
                items:
                  type: object
                  properties:
                    type_url:
                      type: string
                    value:
                      type: string
                      format: byte
      parameters:
        - name: project_did
          in: path
          required: true
          type: string
      tags:
        - Project
  '/projectTxs/{project_did}':
    get:
      operationId: ProjectTx
      description: Get the list of project transactions for a project
      responses:
        '200':
          description: A successful response.
          schema:
            type: object
            properties:
              docs_list:
                type: array
                items:
                  type: object
                  properties:
                    projectDid:
                      type: string
                    recipientDid:
                      type: string
                    amount:
                      type: object
                      properties:
                        denom:
                          type: string
                        amount:
                          type: string
                      description: >-
                        Coin defines a token with a denomination and an
                        amount.


                        NOTE: The amount field is an Int which implements
                        the custom method

                        signatures required by gogoproto.
        default:
          description: An unexpected error response
          schema:
            type: object
            properties:
              error:
                type: string
              code:
                type: integer
                format: int32
              message:
                type: string
              details:
                type: array
                items:
                  type: object
                  properties:
                    type_url:
                      type: string
                    value:
                      type: string
                      format: byte
      parameters:
        - name: project_did
          in: path
          required: true
          type: string
      tags:
        - Project
  /projectParams:
    get:
      operationId: ProjectParams
      description: Get the project module's parameters
      responses:
        '200':
          description: A successful response.
          schema:
            type: object
            properties:
              height:
                type: string
              result:
                type: object
                properties:
                  xco_did:
                    type: string
                  project_minimum_initial_funding:
                    type: array
                    items:
                      type: object
                      properties:
                        denom:
                          type: string
                        amount:
                          type: string
                      description: >-
                        Coin defines a token with a denomination and an amount.


                        NOTE: The amount field is an Int which implements the
                        custom method

                        signatures required by gogoproto.
                  oracle_fee_percentage:
                    type: string
                  node_fee_percentage:
                    type: string
        default:
          description: An unexpected error response
          schema:
            type: object
            properties:
              error:
                type: string
              code:
                type: integer
                format: int32
              message:
                type: string
              details:
                type: array
                items:
                  type: object
                  properties:
                    type_url:
                      type: string
                    value:
                      type: string
                      format: byte
      tags:
        - Project
...